Remove regex, replace with strchr(delim, ch)
Since the regex in `get_key_file()` function doesn't work properly
it is remove for now.
If you want multiple different chars as a delimiter just type them
in sequence, e.g. " \t" for both space and tab.
Note: Multiple char delimiters are not supported that way, but they
weren't before.
A proper way to handle delimiters needs to be added in future.

Signed-off-by: Pascal Arlt <parlt@suse.com>