tests: add test for quoted strings
This test expresses my expectation how double quotes should
be handled. There's no documentation on this subject so far,
it's partly guesswork. But the quote handling should be consistent,
at least. I infer from the current parsing code that double
quotes are supposed to be treated in a special way only if
they occur at the beginning or end of a string. E.g. in

key = "some "quoted" value"

only the first and last double quote is stripped, the other
two are left in place. This is currently the case. But then

key = ""quoted string""

Should be parsed as "quoted string", which is not the case.
Also, I expect quoted leading and trailing whitespace to be
preserved:

key = "  string with whitespace  "

should preserve the leading and trainling blanks.

Moreover, delimiter characters should be parsed correctly at
the beginning of a quoted string:

separator = "========="
arrow = "=>"

Actually this should work without quotes, too, if the set
of delimiters contains at least one non-whitespace character.