Fix typo in README.md: serveral -> several
The debian package linter spotted this typo (in the package description
which was copied from the initial part of the README):

W: libeconf-dev: spelling-error-in-description serveral several
N:
N:    Lintian found a spelling error in the package description. Lintian has a
N:    list of common misspellings that it looks for. It does not have a
N:    dictionary like a spelling checker does. It is particularly picky about
N:    spelling and capitalization in package descriptions since they're very
N:    visible to end users.
N:
N:    Severity: warning
N:
N:    Check: fields/description